Add Offers module with controller, entities, and DTO
Introduced an Offers module that is intended to manage offers within the app. The includes additions are Offers controller leveraging Cron for periodic fetching and validation of offers from providers, Offers entity and DTO defining the structure and validation ofn offer data, and an Offers module combining providers and controller. This allows offers handling and lays groundwork for future provider additions. Also updated App module to incorporate the new Offers module and a Schedule module for managing time-oriented jobs.