Add secondary provider integration for Offers
Integration with a secondary provider has been added to the Offers module. This includes the definition of fetch, validation, and mapping strategies for the provider's response. A transformer class is used to map fields in the response to those expected by the Offers module. An interface describing the payload shape from the secondary provider has also been made. This allows for better abstraction and prepares the ground if there are future provider changes.