Add new ProvidersModule to segregate provider-related functionality
In efforts to make the code more maintainable and clear, provider-related code (secondary and primary providers) have been moved from the offers module to a newly established ProvidersModule. This offers a focused place for all provider-related functionality and simplifies adjustment and navigation of dependencies. The paths of affected files and their dependencies have been accordingly updated.